all added

+++ first_block/real_size.c
@@ -3,16 +3,16 @@
 
 int main() {
     int k;
-    long long int res;
-    long long int another_res;
-    long long int final_res;
+    long int res;
+    long int another_res;
+    long int final_res;
     scanf("%d", &k);
 
     res = pow(2, 30);
     another_res = pow(10, 9);
     final_res = k * (res - another_res);
     
-    printf("%lld", final_res);
+    printf("%ld", final_res);
 
     return 0;
 }
